web/requirements.txt
Flask==0.10.1
requests==2.18.1
python-dateutil==2.5.2
